update: history to core.storage
